+++ packages/depend/__tests__/fixtures/simple.json
