+++ packages/depend/package.json
@@ -4,7 +4,7 @@
   "description": "dependency analyzer and fixer",
   "main": "dist/depend",
   "scripts": {
-    "transpile": "babel src -d dist --ignore coverage,__tests__",
+    "transpile": "babel src -d dist --ignore __tests__",
     "prepublishOnly": "yarn transpile",
     "fmt": "prettier --write '**/*.*'",
     "prettier:diff": "prettier --list-different '**/*.*'",
@@ -21,15 +21,15 @@
     "testEnvironment": "jsdom",
     "coverageDirectory": "<rootDir>/packages/depend/coverage/",
     "collectCoverageFrom": [
-      "**/packages/depend/*.js",
-      "!**/packages/depend/cli.js",
-      "!**/packages/depend/get-packages.js"
+      "**/packages/depend/src/*.js",
+      "!**/packages/depend/src/cli.js",
+      "!**/packages/depend/src/get-packages.js"
     ],
     "transformIgnorePatterns": [
       "node_modules/(?!(react-native|@times-components)/)"
     ],
     "testMatch": [
-      "<rootDir>/packages/depend/src/__tests__/**.test.js"
+      "<rootDir>/packages/depend/__tests__/**.test.js"
     ]
   },
   "repository": {
